Add modal in main.jsp to allow view, change, and deletion of email accounts
Added simple modal in main.jsp to allow view, change, and deletion of email accounts by redirecting users to the specific page and passing the account's informations as well.

The modal is treggered by a click on the account email in the list at the left of the page. There's also some design improvents to list to make it clearer that it's clickable.

Besides that, some scripts and css style definitions were moved to separate files in order to make main.jsp reading and traversal much easier and faster.

Some small additions were mad in order to make it possible to continue development when it's not possible to connect to the database.